Improve Neogit highlights
These changes reduce the shifting of colors that happens as one
navigates hunks with the cursor on a `NeogitStatus` buffer. It reduces
accessibility a tiny bit (by not showing a cursor line highlight inside
hunks) but makes the theme look less broken and makes it easier to
follow what's going on.

It also fixes the `NeogitWinSeparator` highlight, which for some reason
does not use `WinSeparator`.